# GC: optimize sweeping and allocation