The druntime-benchmarks mostly don't show a huge difference for these changes, but rand_small changes from
```
rand_small       3.637 s,    12 MB, 7642 GC 1262 ms, Pauses  444 ms <    0 ms
```
on win64 to
```
rand_small       3.260 s,    12 MB, 7642 GC 1080 ms, Pauses  439 ms <    0 ms
```

Results are more impressive for the extended vdparser test also shown [here](https://github.com/dlang/druntime/pull/2514#issuecomment-475704980):
```
win32: 10.110 s, 1035 MB,   11 GC 1660 ms, Pauses 1361 ms <  562 ms
win64: 10.740 s, 1355 MB,   13 GC 1370 ms, Pauses 1030 ms <  385 ms
```
is changed to
```
win32: 7.600 s, 1035 MB,   11 GC 1594 ms, Pauses 1309 ms <  536 ms
win64: 7.570 s, 1419 MB,   13 GC 1410 ms, Pauses 1078 ms <  417 ms
```