@rainers 2019-03-24
Ouch, sweep and recover cannot be combined if dtors of dead objeccts still access other dead objects. Closing for now...