@Geod24 2019-03-25
> Ouch, sweep and recover cannot be combined if dtors of dead objeccts still access other dead objects. Closing for now...

The dtors are not supposed to access GC allocated objects though, as the order in which they run is not guaranteed. So isn't that exhibiting an issue in current current ?