@rainers 2019-03-25
> The dtors are not supposed to access GC allocated objects though, as the order in which they run is not guaranteed. So isn't that exhibiting an issue in current current ?

It seems to happen in std.regex for reference counting. We could enforce it (by allowing crashes) but it also pretty much restricts even more what can sensibly be done in a dtor. Right now, other dead objects without a dtor are still in a valid state.