@Geod24 2019-03-25
> It seems to happen in std.regex for reference counting.

Shouldn't that be done with `malloc` / `free` instead ?