@rainers 2019-03-25
> Shouldn't that be done with `malloc` / `free` instead ?

Sorry, can't tell. The regex code is indistinguishable from magic. I just noted that a pointer in a dead object was modified by 1.